# Conflicts:
#	lib/Donor/DonorWidgets/organization_card.dart
#	lib/Donor/donor_dashboard.dart
#	lib/Models/Organization.dart
#	lib/Organization/organization_dashboard.dart
#	lib/home_screen.dart
#	lib/login_screen.dart
#	lib/main.dart
#	pubspec.lock
#	pubspec.yaml

+++ lib/Models/Campaign.dart
@@ -3,8 +3,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 class Campaign {
   String title;
   String description;
-  int goalAmount;
-  int amountRaised;
+  double goalAmount;
+  double amountRaised;
   String category;
   Timestamp endDate;
   Timestamp dateCreated;
